third_party/blink/web_tests/external/wpt/fetch/metadata/generated/svg-image.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] sec-fetch-site - HTTPS downgrade-upgrade no attributes
  promise_test: Unhandled rejection with value: object "[object Event]"
[FAIL] sec-fetch-dest no attributes
  assert_array_equals: expected property 0 to be "empty" but got "image" (expected array ["empty"] got ["image"])
[FAIL] sec-fetch-storage-access - Cross-site no attributes
  assert_array_equals: expected property 0 to be "none" but got "active" (expected array ["none"] got ["active"])
Harness: the test ran to completion.
